updated dependencies
