+++ core/pom.xml
@@ -65,7 +65,7 @@
         <dependency>
             <groupId>org.jetbrains.kotlin</groupId>
             <artifactId>kotlin-reflect</artifactId>
-            <version>1.3.72</version>
+            <version>1.4.0-rc</version>
             <scope>provided</scope>
         </dependency>
     </dependencies>
