+++ core/src/test/dependencies/ignoredVersionUpdates.xml
@@ -2,10 +2,12 @@
          xmlns="http://mojo.codehaus.org/versions-maven-plugin/rule/2.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
          xsi:schemaLocation="http://mojo.codehaus.org/versions-maven-plugin/rule/2.0.0 http://mojo.codehaus.org/versions-maven-plugin/xsd/rule-2.0.0.xsd">
     <rules>
+        <!--
         <rule groupId="org.jetbrains.kotlin" comparisonMethod="maven">
             <ignoreVersions>
                 <ignoreVersion>1.4.0-rc</ignoreVersion>
             </ignoreVersions>
         </rule>
+        -->
     </rules>
 </ruleset>
