+++ tests/pom.xml
@@ -45,7 +45,7 @@
         <plugin-update-file-checksum>d25694f9d0674135a9b37b909e84d640</plugin-update-file-checksum>
         <sonar.coverage.jacoco.xmlReportPaths>../coverage/target/site/jacoco-aggregate/jacoco.xml
         </sonar.coverage.jacoco.xmlReportPaths>
-        <kotlin.version>1.3.72</kotlin.version>
+        <kotlin.version>1.4.0-rc</kotlin.version>
     </properties>
 
     <dependencies>
@@ -80,7 +80,7 @@
         <dependency>
             <groupId>org.jetbrains.kotlin</groupId>
             <artifactId>kotlin-reflect</artifactId>
-            <version>1.3.72</version>
+            <version>1.4.0-rc</version>
             <scope>compile</scope>
         </dependency>
         <dependency>
